Changes to CardVertical - Date formatting is done with dayjs library - Routing is done with RouterLink - Resolved the optionality of createdAt and updatedAt fields of Note entity in frontend - Utils related to note moved to infrastructure
